refactor the f32 buf
