+++ crabml-core/src/backends/cpu/buf/mod.rs
@@ -4,5 +4,4 @@ pub mod buf_q8_0;
 
 pub use buf::CpuTensorBuf;
 pub use buf::CpuTensorBufVecDot;
-pub use buf_q8_0::BlockQ8_0;
 pub use buf_q8_0::QuantBufQ8_0;
